MAINT: Refactor more compat checks into compat init
