+++ pandas_datareader/_utils.py
@@ -1,43 +1,12 @@
 import datetime as dt
-from distutils.version import LooseVersion
-
-import pandas as pd
 from pandas import to_datetime
+from requests_file import FileAdapter
+from pandas_datareader.compat import is_number
 
 import requests
-from requests_file import FileAdapter
 import requests_ftp
+
 requests_ftp.monkeypatch_session()
-
-
-if pd.compat.PY3:
-    from urllib.error import HTTPError     # noqa
-else:
-    from urllib2 import HTTPError          # noqa
-
-PANDAS_VERSION = LooseVersion(pd.__version__)
-
-if PANDAS_VERSION >= LooseVersion('0.19.0'):
-    PANDAS_0190 = True
-    from pandas.api.types import is_number                   # noqa
-else:
-    PANDAS_0190 = False
-    from pandas.core.common import is_number                 # noqa
-
-if PANDAS_VERSION >= LooseVersion('0.17.0'):
-    PANDAS_0170 = True
-else:
-    PANDAS_0170 = False
-
-if PANDAS_VERSION >= LooseVersion('0.16.0'):
-    PANDAS_0160 = True
-else:
-    PANDAS_0160 = False
-
-if PANDAS_VERSION >= LooseVersion('0.14.0'):
-    PANDAS_0140 = True
-else:
-    PANDAS_0140 = False
 
 
 class SymbolWarning(UserWarning):
