+++ pandas_datareader/io/sdmx.py
@@ -8,7 +8,7 @@ import pandas as pd
 import pandas.compat as compat
 
 from pandas_datareader.io.util import _read_content
-from pandas_datareader._utils import HTTPError
+from pandas_datareader.compat import HTTPError
 
 
 _STRUCTURE = '{http://www.sdmx.org/resources/sdmxml/schemas/v2_1/structure}'
